io_uring: allow cancelling IORING_OP_URING_CMD
A uring_cmd must either complete quickly or it must be possible to
cancel the command. Otherwise the corresponding resources remain
allocated until the command completes. When the io_uring is released,
these resources include the entire ring.

In the case of waiting on DRM syncobj fences, it is possible for the
waits to never complete.

Signed-off-by: Julian Orth <ju.orth@gmail.com>