drm: add support for io_uring file operations
This patch adds the necessary scaffolding to support the uring_cmd
infrastructure of io_uring. It does not yet add any commands.

uring_cmds are similar to ioctls are identified by a u32. There are no
rules for how these numbers are chosen and they are namespaced to the
file type.

The first-level payload of a uring_cmd is usually stored inside the
struct io_uring_sqe. Therefore no mechanism for having the scaffolding
copy the payload from userspace is required.

Signed-off-by: Julian Orth <ju.orth@gmail.com>